pyproject.toml
[project]
name = "agenzia-ai"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "gunicorn",
    "pydantic",
    "pydantic-settings",
    "supabase",
    "mistralai",
    "twilio",
    "python-multipart",
    "thefuzz",
    "requests",
    "python-dotenv",
    "tenacity",
    "fpdf2",
    "google-auth",
    "langchain-core",
    "langchain-mistralai",
    "langgraph"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501", "PLR2004", "PLC0415"]  # Allow long lines, magic numbers, imports in functions for tests
"upload_data.py" = ["E501", "E701", "PLR1722"]  # Legacy script - allow long lines and inline statements
"**/test_*.py" = ["E501", "PLR2004"]  # Allow long lines and magic numbers in test files
"scripts/**" = ["E501", "PLR2004", "PLC0415", "E722", "F841"]  # Scripts: allow long lines, magic numbers, local imports, bare excepts, unused vars
"simulate_reply.py" = ["PLR2004"]  # Allow magic numbers for HTTP status codes
"demo_*.py" = ["E501"]  # Demo scripts can have long lines

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true
